move variable_cache out of src/debug/variable.rs
